Aggregate cosdata by keys

Aggregatuib function, which basically wraps the dplyr functions group_by and summarise_at so that row-wise aggregation of the cosdata tibble can be made. The operation is defined by a set of functions (given as a list to the .funs variable) and the keys string. The latter speficies the columns, that gives the criterium for the row-wise aggregation.

clump(cosdata, keys = "mm", .funs = mean, opts = coscos::viscos_options())

Arguments

cosdata
The cosdata tibble (see: cook_cosdata).

keys
A string, indicating one or more column-names for data aggregation.

Value

A aggregated tibble

See also

Examples

# aggregate according to month
clump(cosdata, key = "mm", .funs = mean, opts = coscos::viscos_options())

#> # A tibble: 12 x 5
#>       mm QOBS_0001 QSIM_0001 QOBS_0002 QSIM_0002
#>    <int>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1     1      3.87      3.38      5.06      4.56
#>  2     2      3.07      2.91      4.11      3.72
#>  3     3      3.52      3.26      7.73      6.45
#>  4     4      7.94     10.2      18.0      14.7 
#>  5     5     16.1      19.4      11.2      12.3 
#>  6     6     15.4      17.3      11.6      11.9 
#>  7     7     13.3      11.7      10.8      11.7 
#>  8     8      8.46      7.93      7.62      7.40
#>  9     9     10.3       9.00     10.9       9.90
#> 10    10      6.98      6.20      6.36      5.02
#> 11    11      5.51      5.02      6.48      5.52
#> 12    12      4.81      3.61      6.18      4.36

# aggregate according to year and month
clump(cosdata, key = c("yyyy","mm"), .funs = mean, opts = coscos::viscos_options())

#> # A tibble: 36 x 6
#> # Groups:   yyyy [?]
#>     yyyy    mm QOBS_0001 QSIM_0001 QOBS_0002 QSIM_0002
#>    <int> <int>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1  2007     1      3.55      4.05      6.08      6.61
#>  2  2007     2      2.99      3.03      3.95      4.17
#>  3  2007     3      3.25      3.94      9.05      6.97
#>  4  2007     4      8.06     11.8       8.72      8.31
#>  5  2007     5     11.6      12.5       7.21      8.68
#>  6  2007     6     10.7      13.0       4.14      8.73
#>  7  2007     7     14.3      12.1       3.44      5.54
#>  8  2007     8      7.24      7.13      3.15      4.55
#>  9  2007     9     13.6      10.4      19.1      17.5 
#> 10  2007    10      8.13      5.83      7.13      5.32
#> # ... with 26 more rows