Forge cosdata

forge allows to filter, tidy or re-arrange cosdata into new formats.

forge(cosdata, form = "tidy", opts = coscos::viscos_options())

Arguments

cosdata
The strictly defined data format (cosdata) used within viscos (see: cook_cosdata)

form
A string that defines into which form the cosdata is "forged" (see: 'details')

opts
The options for viscos (see: viscos_options)

Details

The forge function transforms cosdata into a number of different forms. Currently the following forms (strings) can be used:

tidy ... Transfroms the cosdata into a tidy form.

s_data ... Filters the \(o\) columns out of cosdata.

o_data ... Filters the \(s\) columns out of cosdata.

s_only ... Extracts the \(s\) columns only.

o_only ... Extracts the \(o\) columns only.

shell ... Returns the cosdata withouth the data columns(i.e. the \(o & s\) columns).

See also

Examples

# tidy: 
forge( coscos::viscos_example() )

#> Error in gather_(., key_col = c("key"), value_col = c("value"), gather_cols = c(le_q_names)): konnte Funktion "gather_" nicht finden

# only s-data columns:
forge( coscos::viscos_example(), form = "s_only" )

#> # A tibble: 26,304 x 2
#>    QSIM_0001 QSIM_0002
#>        <dbl>     <dbl>
#>  1      3.48      3.11
#>  2      3.48      3.11
#>  3      3.48      3.11
#>  4      3.48      3.11
#>  5      3.48      3.11
#>  6      3.49      3.11
#>  7      3.49      3.11
#>  8      3.49      3.11
#>  9      3.49      3.11
#> 10      3.50      3.11
#> # ... with 26,294 more rows