mould cosdata

mould allows to filter, tidy or re-arrange cosdata into new formats.

mould(cosdata, form = "tidy", opts = coscos::viscos_options())

Arguments

cosdata
The strictly defined data format (cosdata) used within viscos (see: cook_cosdata)

form
A string that defines into which form the cosdata is "mouldd" (see: 'details')

opts
The options for viscos (see: viscos_options)

Details

The mould function transforms cosdata into a number of different forms. Currently the following forms (strings) can be used:

tidy ... Transfroms the cosdata into a tidy form.

s_data ... Filters the \(o\) columns out of cosdata.

o_data ... Filters the \(s\) columns out of cosdata.

s_only ... Extracts the \(s\) columns only.

o_only ... Extracts the \(o\) columns only.

shell ... Returns the cosdata withouth the data columns(i.e. the \(o & s\) columns).

See also

Examples

# tidy: 
mould( coscos::viscos_example() )

#> # A tibble: 105,216 x 4
#>    posixdate           period key       value
#>    <dttm>               <int> <chr>     <dbl>
#>  1 2007-01-01 00:00:00      0 QOBS_0001  2.98
#>  2 2007-01-01 01:00:00      0 QOBS_0001  2.89
#>  3 2007-01-01 02:00:00      0 QOBS_0001  2.64
#>  4 2007-01-01 03:00:00      0 QOBS_0001  2.51
#>  5 2007-01-01 04:00:00      0 QOBS_0001  2.42
#>  6 2007-01-01 05:00:00      0 QOBS_0001  2.34
#>  7 2007-01-01 06:00:00      0 QOBS_0001  2.32
#>  8 2007-01-01 07:00:00      0 QOBS_0001  2.31
#>  9 2007-01-01 08:00:00      0 QOBS_0001  2.27
#> 10 2007-01-01 09:00:00      0 QOBS_0001  2.25
#> # ... with 105,206 more rows

# only s-data columns:
mould( coscos::viscos_example(), form = "s_only" )

#> # A tibble: 26,304 x 2
#>    QSIM_0001 QSIM_0002
#>        <dbl>     <dbl>
#>  1      3.48      3.11
#>  2      3.48      3.11
#>  3      3.48      3.11
#>  4      3.48      3.11
#>  5      3.48      3.11
#>  6      3.49      3.11
#>  7      3.49      3.11
#>  8      3.49      3.11
#>  9      3.49      3.11
#> 10      3.50      3.11
#> # ... with 26,294 more rows