Aggregate or clump by keys

An aggregate function, which wraps the dplyr functions group_by and summarise_at for a row-wise summary of the cosdata tibble. It is defined by the set of functions given by .funs variable, and the keys string. The latter specifies the columns, according to whose unique entries the row-wise aggregation is conducted.

clump(cosdata, keys = "mm", .funs = mean, opts = coscos::viscos_options())

Arguments

cosdata
The cosdata tibble (see: cook_cosdata).

keys
A string, indicating one or more column-names for data aggregation.

.funs
A aggregation function or list of aggregation functions.

Value

An aggregated tibble.

See also

Examples

# aggregate according to month
cosdata <- viscos_example()
clump(cosdata, key = "mm", .funs = mean)

#> # A tibble: 12 x 5
#>       mm QOBS_0001 QSIM_0001 QOBS_0002 QSIM_0002
#>    <int>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1     1      3.87      3.38      5.06      4.56
#>  2     2      3.07      2.91      4.11      3.72
#>  3     3      3.52      3.26      7.73      6.45
#>  4     4      7.94     10.2      18.0      14.7 
#>  5     5     16.1      19.4      11.2      12.3 
#>  6     6     15.4      17.3      11.6      11.9 
#>  7     7     13.3      11.7      10.8      11.7 
#>  8     8      8.46      7.93      7.62      7.40
#>  9     9     10.3       9.00     10.9       9.90
#> 10    10      6.98      6.20      6.36      5.02
#> 11    11      5.51      5.02      6.48      5.52
#> 12    12      4.81      3.61      6.18      4.36

# aggregate according to year and month
cosdata <- viscos_example()
clump(cosdata, key = c("yyyy", "mm"), .funs = median)

#> # A tibble: 36 x 6
#> # Groups:   yyyy [?]
#>     yyyy    mm QOBS_0001 QSIM_0001 QOBS_0002 QSIM_0002
#>    <int> <int>     <dbl>     <dbl>     <dbl>     <dbl>
#>  1  2007     1      2.71      3.50      4.15      4.47
#>  2  2007     2      2.55      3.03      3.80      4.05
#>  3  2007     3      2.79      3.50      8.47      6.79
#>  4  2007     4      8.84     13.0       8.98      8.41
#>  5  2007     5     11.6      11.7       5.51      7.50
#>  6  2007     6     10.4      11.3       4.01      7.77
#>  7  2007     7     13.4       9.83      3.06      5.07
#>  8  2007     8      6.86      6.06      2.81      4.24
#>  9  2007     9     13.5       9.49      9.80     10.6 
#> 10  2007    10      7.27      5.38      4.39      3.91
#> # ... with 26 more rows

# aggregate with two functions according to year 
cosdata <- viscos_example()
clump( cosdata, key = "yyyy", .funs = list(a = mean, b = median) )

#> # A tibble: 3 x 9
#>    yyyy QOBS_0001_a QSIM_0001_a QOBS_0002_a QSIM_0002_a QOBS_0001_b QSIM_0001_b
#>   <int>       <dbl>       <dbl>       <dbl>       <dbl>       <dbl>       <dbl>
#> 1  2007        7.93        7.65        7.41        7.27        6.55        5.61
#> 2  2008        7.56        8.77        8.24        7.95        5.59        5.25
#> 3  2009        9.39        8.59       10.9         9.18        6.82        5.52
#> # ... with 2 more variables: QOBS_0002_b <dbl>, QSIM_0002_b <dbl>